django/requirements.txt
django==2.1
requests>=2.5.2
PyJWT>=1.6.4
